refactored Telegram.js; moved generating of UI buttons into another class; Updated ToDo;
